Module rusty_saw_view::json_structs

json_structs is a collection of submodules that define structures that map to the JSON data found at various endpoints for Hyperledger Sawtooth.

As of now there are only modules that represent the /blocks and /state endpoints.

Modules

json_blocks
json_blocks is a collection of structures that represent the JSON structured data from the /blocks endpoint of Hyperledger Sawtooth. Also includes methods for retrieving useful information and displaying these structures.

json_state
json_state is a collection of structures that represent the JSON structured data from the /state endpoint of Hyperledger Sawtooth. Also includes methods for retrieving useful information and displaying these structures.

Functions

get_partial_string
Returns a string of the first n and last m characters of the given string