Module rusty_saw_view::json_structs::json_state

json_state is a collection of structures that represent the JSON structured data from the /state endpoint of Hyperledger Sawtooth. Also includes methods for retrieving useful information and displaying these structures.

Typically when you parse JSON data from the /state endpoint you'll store it in the root StateData structure. That structure contains methods for printing out the contents of the blockchain. Use display_full_data when printing to the terminal, and display_full_data_no_color if you want to pipe the output to a file.

Structs

State
A structure that represents a state found at the /state endpoint. States contain an address and the data stored at that address

StateData
A structure that represents the root data item found at the /state endpoint.